Add Encore::Persister and move serializer into Encore::Serializer
