+++ lib/encore.rb
@@ -1,4 +1,7 @@
-require 'encore/base'
+require 'encore/persister/base'
+require 'encore/serializer/base'
+
+require 'encore/version'
 
 module Encore
 end
