--- lib/encore/linked_loading/has_many.rb
@@ -1,11 +1,0 @@
-module Encore
-  module LinkedLoading
-    module HasMany
-      extend ActiveSupport::Concern
-
-      def many(item, reflection)
-        item.send("#{reflection.name.to_s.singularize}_ids").map(&:to_s)
-      end
-    end
-  end
-end
