+++ lib/encore/serializer/eager_loading.rb
@@ -1,0 +1,11 @@
+module Encore
+  module Serializer
+    module EagerLoading
+      extend ActiveSupport::Concern
+
+      def add_eager_loading(collection, includes)
+        collection.includes(includes)
+      end
+    end
+  end
+end
