+++ lib/encore/serializer/linked_loading/has_many.rb
@@ -1,0 +1,13 @@
+module Encore
+  module Serializer
+    module LinkedLoading
+      module HasMany
+        extend ActiveSupport::Concern
+
+        def many(item, reflection)
+          item.send("#{reflection.name.to_s.singularize}_ids").map(&:to_s)
+        end
+      end
+    end
+  end
+end
