+++ lib/encore/serializer/main_paging.rb
@@ -1,0 +1,13 @@
+module Encore
+  module Serializer
+    module MainPaging
+      extend ActiveSupport::Concern
+
+      def add_main_pagination(collection)
+        {
+          serializer.root_key => pagination_for(collection)
+        }
+      end
+    end
+  end
+end
