+++ lib/encore/serializer/main_serialized.rb
@@ -1,0 +1,13 @@
+module Encore
+  module Serializer
+    module MainSerialized
+      extend ActiveSupport::Concern
+
+      def add_main_serialized(collection)
+        {
+          serializer.root_key => collection.map { |o| serializer.new(o).as_json }
+        }
+      end
+    end
+  end
+end
